build_utils/pyproject.toml
[project]
name = "build-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "scipy>=1.16.1",
]
